+++ __tests__/hooks/auth/useCreateMailToken.test.ts
@@ -2,7 +2,8 @@ import useCreateMailToken from '@hooks/auth/useCreateMailToken';
 import { renderHook } from '@testing-library/react-hooks';
 
 import { useAuthRepository } from 'src/commons/di/containerContext';
-import { IAuthRepository } from 'src/commons/network/types/interfaces';
+import IAuthRepository from 'src/commons/network/repositories/AuthRepository/AuthRepository.interface';
+
 import { anything, instance, mock, reset, verify, when } from 'ts-mockito';
 
 jest.mock('src/commons/di/containerContext', () => {
