+++ __tests__/hooks/auth/useVerifyMailToken.test.ts
@@ -1,9 +1,9 @@
-import useCreateMailToken from '@hooks/auth/useCreateMailToken';
 import useVerifyMailToken from '@hooks/auth/useVerifyMailToken';
 import { renderHook } from '@testing-library/react-hooks';
 
 import { useAuthRepository } from 'src/commons/di/containerContext';
-import { IAuthRepository } from 'src/commons/network/types/interfaces';
+import IAuthRepository from 'src/commons/network/repositories/AuthRepository/AuthRepository.interface';
+
 import { anything, instance, mock, reset, verify, when } from 'ts-mockito';
 
 jest.mock('src/commons/di/containerContext', () => {
