+++ __tests__/repositories/AuthRepository.test.ts
@@ -1,8 +1,9 @@
 import ERROR_MESSAGE from '@constants/errorMessage';
 import { IUser } from 'src/commons/graphql-types/generated/types';
-import AuthRepository from 'src/commons/network/repositories/AuthRepository';
+import AuthRepository from 'src/commons/network/repositories/AuthRepository/AuthRepository';
+import IAuthRepository from 'src/commons/network/repositories/AuthRepository/AuthRepository.interface';
 import IHttpClient from 'src/commons/network/types/HttpClient.interface';
-import { IAuthRepository } from 'src/commons/network/types/interfaces';
+
 import { anything, instance, mock, reset, when } from 'ts-mockito';
 
 describe('AuthRepository', () => {
