+++ src/commons/di/containerContext.ts
@@ -1,9 +1,9 @@
 import constate from 'constate';
 import { useMemo } from 'react';
 import ApolloHttpClient from '../network/ApolloHttpClient';
-import AuthRepository from '../network/repositories/AuthRepository';
+import AuthRepository from '../network/repositories/AuthRepository/AuthRepository';
 import IHttpClient from '../network/types/HttpClient.interface';
-import { IAuthRepository } from '../network/types/interfaces';
+import { IAuthRepository } from '../network/repositories/AuthRepository/AuthRepository.interface';
 
 const useContainerContext = () => {
   const httpClient: IHttpClient = useMemo(() => {
